Added certificates and tool to generate passwords
cacert.pem is certificate of certificate authority while browserCert.pfx
is certificate for browser issued by CA. This certificates are not
secure, they are used only for academic purposes. Secure certificates
can be retrieved by trusted certificate authorities. The folder
"password generator" is a simple tool to generate password with salt for
user registration in db.